Set up a new files layout
